backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
lightrag-hku>=1.0.0
groq>=0.11.0
python-dotenv>=1.0.0
numpy>=1.26.0
